HSEARCH-3282 Move HibernateOrmSearchScrollableResultsAdapter to SPI
So that we can use it in the migration helper
